test_deploy/requirements.txt
sk==2.3.2
transformers==4.32.0
pytorch==2.1.0
pandas==1.5.3
pdfplumber==0.7.4
python-pptx==0.6.21
pytesseract==0.3.10
Pillow==9.6.1
google-api-python-client==2.91.0
google-auth==2.15.0
sqlite3

